Extract out shake detection logic
This should improve test coverage
